Falcor 4.2

+++ Source/Falcor/Utils/AlignedAllocator.h
@@ -13,7 +13,7 @@
  #    contributors may be used to endorse or promote products derived
  #    from this software without specific prior written permission.
  #
- # THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS ``AS IS'' AND ANY
+ # THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS "AS IS" AND ANY
  # EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  # IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
  # PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL THE COPYRIGHT OWNER OR
@@ -148,8 +148,8 @@ namespace Falcor
             return &*iter;
         }
 
-        int mMinAlignment = 16;
-        int mCacheLineSize = 128;
+        size_t mMinAlignment = 16;
+        size_t mCacheLineSize = 128;
         std::vector<uint8_t> mBuffer;
     };
 }
